Docker/websocket-server/requirements.txt
aiohttp
python-socketio
sympy
numpy
matplotlib